Ultra-rough POC for JSMpeg
In this state, Hypcast is able to successfully stream MPEG1 + MP2 live
video over a websocket using JSMpeg. While the implementation is a hack
and quite buggy, the basic functionality does work.

Unfortunately, I'm very unimpressed with the quality of the JSMpeg
stream. I've tried tweaking a few player dials, but the audio is
extremely laggy and/or choppy (at least at higher bitrates, but I
consider that a requirement). I've also noticed that the stream can
still take ~ 5 - 10 seconds to start, which isn't much better than the
HLS implementation in the first place.

I'll still be considering alternative implementations for truly-live
streaming, e.g. WebRTC. However, this would be fairly complex and likely
require a custom C++ server.