Create initial working GStreamer prototype
This code is able to construct a valid GStreamer pipeline that produces
HLS segments and a playlist file. Tested manually, since it's currently
not connected to anything.

Took a lot of work to get a working GStreamer installation (at least,
one new enough for my purposes), but the actual pipeline is a lot
simpler since it's no longer going through multiple child processes.