Bump to v1.0.0; end of v1 development?
After more than 4 years of development and maintenance, I think it's
reasonable to say that "v0.0.1" isn't a proper version moniker for
Hypcast.

However, the main reason I'm comfortable bumping to v1.0.0 is that I
don't expect any major changes to this version of Hypcast. It works well
enough for how it was designed.

That said, "how it was designed" could potentially be improved. In the
past I've considered making two improvements to Hypcast: Using GStreamer
to handle transforming and encoding the input TV signal, and using
WebRTC to send it to clients. After experimenting with both technologies
(particularly GStreamer) in fits and starts over the past year or two, I
think I have a better idea of how to make this happen for real, and I
want to iterate on that. This new version of Hypcast has the potential
to be much more performant, and much closer to the experience of
watching a real television (vs. a 10 second delay every time you change
channels).

As such, I might not touch this branch so heavily anymore.